jarvis/frontend/requirements.txt
sounddevice>=0.5.1
numpy>=2.0.0
websockets>=12.0
python-dotenv>=1.0.1
openwakeword>=0.6.0
webrtcvad>=2.0.10
